+++ src/main/java/it/pagopa/pn/bff/mappers/notifications/NotificationAarQrCodeMapper.java
@@ -7,6 +7,10 @@ import it.pagopa.pn.bff.generated.openapi.server.v1.dto.BffCheckAarResponse;
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
 
+/**
+ * Mapstruct mapper interface, used to map the parameters (from bff version to ms version and viceversa)
+ * used to call the AArQrCodeCheck api
+ */
 @Mapper
 public interface NotificationAarQrCodeMapper {
     // Instance of the mapper
